Remove `apt-get upgrade` and condense installs
I just spent the better part of two hours debugging this:

```
» psql -h 0.0.0.0 -p 7432
psql: server closed the connection unexpectedly
        This probably means the server terminated abnormally
        before or while processing the request.
```

when trying to connect to the running container.

I was not able to reproduce the error when using the `postgres:9.6.5`
image directly, thus something in the build process must be severing the
connection when I try to connect.

The only explanation I can muster is that the `apt-get upgrade`
accidentally upgrades PostgreSQL into a silently-incompatible place.

Generally speaking, doing an `apt-get upgrade` in a docker build is an
anti-pattern, as you should just use a more recent upstream image
instead. So I removed it. After that, everything seems to work.

Also, I merged the `apt-get install` commands into a single place, for
purposes of reducing the overall size and layer count of the built
image.